# aspnetcore/src/IO.Swagger/Controllers/DronesApi.cs
/*
 * PackagesOfFuture.Api
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class DronesApiController : ControllerBase
    { 
        /// <summary>
        /// Moves drone to different sorting
        /// </summary>
        /// <param name="droneId">Id of drone</param>
        /// <param name="body">Details of moving drone</param>
        /// <response code="200">Success</response>
        /// <response code="204">When drone was moved</response>
        /// <response code="400">When error regarding input occurred</response>
        [HttpPost]
        [Route("/Drones/{droneId}/moveToSorting")]
        [ValidateModelState]
        [SwaggerOperation("DronesDroneIdMoveToSortingPost")]
        public virtual IActionResult DronesDroneIdMoveToSortingPost([FromRoute][Required]int? droneId, [FromBody]MoveDroneToSortingDto body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            //TODO: Uncomment the next line to return response 204 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(204);

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            throw new NotImplementedException();
        }

        /// <summary>
        /// Gets all drones
        /// </summary>
        /// <response code="200">When there are drones</response>
        /// <response code="404">If there are no drones</response>
        [HttpGet]
        [Route("/Drones")]
        [ValidateModelState]
        [SwaggerOperation("DronesGet")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<DroneDto>), description: "When there are drones")]
        public virtual IActionResult DronesGet()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<DroneDto>));

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "[ {\n  \"range\" : 0,\n  \"model\" : \"model\"\n}, {\n  \"range\" : 0,\n  \"model\" : \"model\"\n} ]";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<DroneDto>>(exampleJson)
                        : default(List<DroneDto>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Registers a new drone
        /// </summary>
        /// <param name="body">Representation of drone to register</param>
        /// <response code="200">Success</response>
        /// <response code="204">When drone was registered</response>
        /// <response code="400">When error regarding input occurred</response>
        [HttpPost]
        [Route("/Drones")]
        [ValidateModelState]
        [SwaggerOperation("DronesPost")]
        public virtual IActionResult DronesPost([FromBody]RegisterDroneDto body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            //TODO: Uncomment the next line to return response 204 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(204);

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            throw new NotImplementedException();
        }
    }
}