# aspnetcore/src/IO.Swagger/Controllers/VehiclesApi.cs
/*
 * PackagesOfFuture.Api
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class VehiclesApiController : ControllerBase
    { 
        /// <summary>
        /// Gets all vehicles
        /// </summary>
        /// <response code="200">When there are vehicles</response>
        /// <response code="404">If there are no vehicles</response>
        [HttpGet]
        [Route("/Vehicles")]
        [ValidateModelState]
        [SwaggerOperation("VehiclesGet")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<VehicleDto>), description: "When there are vehicles")]
        public virtual IActionResult VehiclesGet()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<VehicleDto>));

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "[ {\n  \"model\" : \"model\",\n  \"id\" : 0\n}, {\n  \"model\" : \"model\",\n  \"id\" : 0\n} ]";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<VehicleDto>>(exampleJson)
                        : default(List<VehicleDto>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Registers a new vehicle
        /// </summary>
        /// <param name="body">Representation of vehicle to register</param>
        /// <response code="200">Success</response>
        /// <response code="204">When vehicle was registered</response>
        /// <response code="400">When error regarding input occurred</response>
        [HttpPost]
        [Route("/Vehicles")]
        [ValidateModelState]
        [SwaggerOperation("VehiclesPost")]
        public virtual IActionResult VehiclesPost([FromBody]RegisterVehicleDto body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            //TODO: Uncomment the next line to return response 204 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(204);

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            throw new NotImplementedException();
        }
    }
}