# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "PackagesOfFuture.Api",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/Drones" : {
      "get" : {
        "tags" : [ "Drones" ],
        "summary" : "Gets all drones",
        "responses" : {
          "200" : {
            "description" : "When there are drones",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/DroneDto"
                  },
                  "x-content-type" : "text/plain"
                }
              },
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/DroneDto"
                  }
                }
              },
              "text/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/DroneDto"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "If there are no drones"
          }
        }
      },
      "post" : {
        "tags" : [ "Drones" ],
        "summary" : "Registers a new drone",
        "requestBody" : {
          "description" : "Representation of drone to register",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterDroneDto"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterDroneDto"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterDroneDto"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          },
          "204" : {
            "description" : "When drone was registered"
          },
          "400" : {
            "description" : "When error regarding input occurred"
          }
        }
      }
    },
    "/Drones/{droneId}/moveToSorting" : {
      "post" : {
        "tags" : [ "Drones" ],
        "summary" : "Moves drone to different sorting",
        "parameters" : [ {
          "name" : "droneId",
          "in" : "path",
          "description" : "Id of drone",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "description" : "Id of drone",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "description" : "Details of moving drone",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MoveDroneToSortingDto"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MoveDroneToSortingDto"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MoveDroneToSortingDto"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          },
          "204" : {
            "description" : "When drone was moved"
          },
          "400" : {
            "description" : "When error regarding input occurred"
          }
        }
      }
    },
...
    "/Packages" : {
      "get" : {
        "tags" : [ "Packages" ],
        "summary" : "Gets all packages",
        "responses" : {
          "200" : {
            "description" : "When there are packages",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/PackageDto"
                  },
                  "x-content-type" : "text/plain"
                }
              },
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/PackageDto"
                  }
                }
              },
              "text/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/PackageDto"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "If there are no packages"
          }
        }
      },
      "post" : {
        "tags" : [ "Packages" ],
        "summary" : "Registers a new package",
        "requestBody" : {
          "description" : "Representation of package to register",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterPackageDto"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterPackageDto"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterPackageDto"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          },
          "204" : {
            "description" : "When package was registered"
          },
          "400" : {
            "description" : "When error regarding input occurred"
          }
        }
      }
    },
...
    "/Vehicles" : {
      "get" : {
        "tags" : [ "Vehicles" ],
        "summary" : "Gets all vehicles",
        "responses" : {
          "200" : {
            "description" : "When there are vehicles",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/VehicleDto"
                  },
                  "x-content-type" : "text/plain"
                }
              },
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/VehicleDto"
                  }
                }
              },
              "text/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/VehicleDto"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "If there are no vehicles"
          }
        }
      },
      "post" : {
        "tags" : [ "Vehicles" ],
        "summary" : "Registers a new vehicle",
        "requestBody" : {
          "description" : "Representation of vehicle to register",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterVehicleDto"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterVehicleDto"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterVehicleDto"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          },
          "204" : {
            "description" : "When vehicle was registered"
          },
          "400" : {
            "description" : "When error regarding input occurred"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "AddressDto" : {
        "type" : "object",
        "properties" : {
          "street" : {
            "type" : "string",
            "nullable" : true
          },
          "houseAndFlatNumber" : {
            "type" : "string",
            "nullable" : true
          },
          "city" : {
            "type" : "string",
            "nullable" : true
          },
          "postalCode" : {
            "type" : "string",
            "nullable" : true
          }
        },
        "additionalProperties" : false,
        "example" : {
          "houseAndFlatNumber" : "houseAndFlatNumber",
          "city" : "city",
          "street" : "street",
          "postalCode" : "postalCode"
        }
      },
...
      "DroneDto" : {
        "type" : "object",
        "properties" : {
          "model" : {
            "type" : "string",
            "nullable" : true
          },
          "range" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "additionalProperties" : false,
        "example" : {
          "range" : 0,
          "model" : "model"
        }
      },
      "RegisterDroneDto" : {
        "type" : "object",
        "properties" : {
          "model" : {
            "type" : "string",
            "nullable" : true
          },
          "range" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "additionalProperties" : false
      },
      "MoveDroneToSortingDto" : {
        "type" : "object",
        "properties" : {
          "droneId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sortingId" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "additionalProperties" : false
      },
...
      "PackageStatus" : {
        "type" : "integer",
        "format" : "int32",
        "enum" : [ 0, 1, 2 ]
      },
      "PaymentType" : {
        "type" : "integer",
        "format" : "int32",
        "enum" : [ 1, 2, 3, 4 ]
      },
      "PaymentStatus" : {
        "type" : "integer",
        "format" : "int32",
        "enum" : [ 0, 1, 2 ]
      },
      "PaymentDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "type" : {
            "$ref" : "#/components/schemas/PaymentType"
          },
          "amount" : {
            "type" : "number",
            "format" : "double"
          },
          "status" : {
            "$ref" : "#/components/schemas/PaymentStatus"
          }
        },
        "additionalProperties" : false,
        "example" : {
          "amount" : 9.301444243932576,
          "id" : 2,
          "type" : 7,
          "status" : 3
        }
      },
      "PackageDto" : {
        "type" : "object",
        "properties" : {
          "deliveryDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "$ref" : "#/components/schemas/PackageStatus"
          },
          "width" : {
            "type" : "number",
            "format" : "double"
          },
          "height" : {
            "type" : "number",
            "format" : "double"
          },
          "length" : {
            "type" : "number",
            "format" : "double"
          },
          "weight" : {
            "type" : "number",
            "format" : "double"
          },
          "payment" : {
            "$ref" : "#/components/schemas/PaymentDto"
          },
          "deliveryAddress" : {
            "$ref" : "#/components/schemas/AddressDto"
          },
          "receiveAddress" : {
            "$ref" : "#/components/schemas/AddressDto"
          },
          "deliveryAddressId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "receiveAddressId" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "additionalProperties" : false,
        "example" : {
          "deliveryAddressId" : 2,
          "receiveAddressId" : 4,
          "deliveryAddress" : {
            "houseAndFlatNumber" : "houseAndFlatNumber",
            "city" : "city",
            "street" : "street",
            "postalCode" : "postalCode"
          },
          "receiveAddress" : null,
          "width" : 6.027456183070403,
          "length" : 5.962133916683182,
          "weight" : 5.637376656633329,
          "payment" : {
            "amount" : 9.301444243932576,
            "id" : 2,
            "type" : 7,
            "status" : 3
          },
          "deliveryDate" : "2000-01-23T04:56:07.000+00:00",
          "status" : 0,
          "height" : 1.4658129805029452
        }
      },
      "PackageDetailsDto" : {
        "type" : "object",
        "properties" : {
          "width" : {
            "type" : "number",
            "format" : "double"
          },
          "height" : {
            "type" : "number",
            "format" : "double"
          },
          "length" : {
            "type" : "number",
            "format" : "double"
          },
          "weight" : {
            "type" : "number",
            "format" : "double"
          }
        },
        "additionalProperties" : false
      },
      "CreatePaymentDto" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "$ref" : "#/components/schemas/PaymentType"
          },
          "amount" : {
            "type" : "number",
            "format" : "double"
          }
        },
        "additionalProperties" : false
      },
      "RegisterPackageDto" : {
        "type" : "object",
        "properties" : {
          "deliveryAddress" : {
            "$ref" : "#/components/schemas/AddressDto"
          },
          "receiveAddress" : {
            "$ref" : "#/components/schemas/AddressDto"
          },
          "package" : {
            "$ref" : "#/components/schemas/PackageDetailsDto"
          },
          "serviceId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "payment" : {
            "$ref" : "#/components/schemas/CreatePaymentDto"
          }
        },
        "additionalProperties" : false
      },
...
      "VehicleDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "model" : {
            "type" : "string",
            "nullable" : true
          }
        },
        "additionalProperties" : false,
        "example" : {
          "model" : "model",
          "id" : 0
        }
      },
      "RegisterVehicleDto" : {
        "type" : "object",
        "properties" : {
          "model" : {
            "type" : "string",
            "nullable" : true
          }
        },
        "additionalProperties" : false
      }
    }
  }
}